pyproject.toml
[tool.poetry]
name = "hw14"
version = "0.1.0"
description = ""
authors = ["taras55001 <taras5500@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.100.0"}
uvicorn = {extras = ["standard"], version = "^0.23.1"}
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
asyncpg = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-jwt-auth = "^0.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
fastapi-mail = "^1.4.1"
redis = "^4.6.0"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.33.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.0"
openai = "^0.27.8"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"
aiosqlite = "^0.19.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
